refactor(frontend): padroniza estrutura de pastas, arquivos e atualiza imports
- Renomeia pastas para camelCase (ex: auth, profile, detail)
- Renomeia componentes e telas para PascalCase
- Centraliza componentes aninhados no diretório global components/
- Atualiza caminhos de importação em todo o projeto
