+++ src/day-counter/navigation/AuthRoutes.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import AuthScreen from "../screens/login-cadastro/AuthScreen";
+import AuthScreen from "../screens/auth/AuthScreen";
 
 const Stack = createNativeStackNavigator();
