+++ src/day-counter/screens/profile/profileStyles.js
